New file apple_base.py (~130–140 lines):
The 4 response dataclasses verbatim (preserving all docs/comments/style)
_flatten_messages() verbatim
_AppleBaseLM(BaseLM) abstract base with a single concrete method: a unified _build_response(text, usage=None) that both adapters can inherit
apple_fm.py edits:

Remove the 4 dataclasses + _flatten_messages (~120 lines)
Import from apple_base instead
AppleFoundationLM(BaseLM) → AppleFoundationLM(_AppleBaseLM)
Drop its _build_response override (now inherited)
Result: 530 lines
apple_local.py edits:

Swap from dspy.clients.apple_fm import ... → from dspy.clients.apple_base import ...
Remove inline from dspy.clients.apple_fm import _flatten_messages inside _apply_chat_template
AppleLocalLM(BaseLM) → AppleLocalLM(_AppleBaseLM)
Drop its _build_response override (now inherited)
Switch _LocalStreamChunk from @dataclass shorthand to @dataclasses.dataclass to match file conventions
Result: 710 lines
